Add files via upload
heater on for 1 hour duration functionality now works.  Used an if statement in the Blynk V8 pin status call, calling the appropriate functions.  It took a little while to get the kinks worked out but seems to be working well.

Future improvements - add date and time functionality so that a schedule can be built.